requirements.txt
requests
fastapi
uvicorn[standard]
python-dotenv
python-multipart
replicate